move ThemeProvider and ThemeToggle components
